move draft tools
