+++ ydb/tools/mnc/cli/commands/agent.py
@@ -6,8 +6,8 @@ import shlex
 import rich
 import yaml
 
-from ydb.tools.mnc.lib import agent_client, common, deploy_ctx, progress, service
-from ydb.tools.mnc.lib.draft import term, tools
+from ydb.tools.mnc.lib import agent_client, common, deploy_ctx, progress, service, tools
+from ydb.tools.mnc.lib.draft import term
 from ydb.tools.mnc.scheme import agent
